+++ quickshell/popups/settings/BarSection.qml
@@ -21,7 +21,7 @@ Item {
     anchors.left: parent.left
     anchors.right: parent.right
     anchors.top: parent.top
-    spacing: 12
+    spacing: 16
 
     SetCard {
       sys: root.sys